refactor : UI 디렉토리 하위 구조 변경
