--- frontend/src/shared/ui/button/standard-button/index.ts
@@ -1,1 +1,0 @@
-export { default as Button } from './ui';
