--- frontend/src/shared/ui/calendar/index.ts
@@ -1,1 +1,0 @@
-export { MonthlyCalendar } from './monthly-calendar';
