--- frontend/src/shared/ui/calendar/monthly-calendar/index.ts
@@ -1,1 +1,0 @@
-export { default as MonthlyCalendar } from './ui';
