+++ frontend/src/shared/ui/index.ts
@@ -1,3 +1,3 @@
-export { TextInput } from './input';
-export { Button } from './button/standard-button';
-export { MonthlyCalendar } from './calendar/monthly-calendar';
+export { default as StandardButton } from './button/StandardButton';
+export { default as MonthlyCalendar } from './calendar/MonthlyCalendar';
+export { default as TextInput } from './input/TextInput';
