--- frontend/src/shared/ui/input/index.ts
@@ -1,1 +1,0 @@
-export { TextInput } from './text-input';
