--- frontend/src/shared/ui/input/text-input/index.ts
@@ -1,1 +1,0 @@
-export { default as TextInput } from './ui';
