+++ frontend/src/shared/ui/input/types.ts
